Move RSC helpers out of aggregate parser
